### Overview

Implement the FF3-1 cipher with unsigned 128-bit integers, a widely supported extension type on modern 64-bit compilers.

This PR also introduces a simple set of benchmarks and an associated Makefile target.

### Build

These changes set the 128-bit integer mode to be the default build mode where supported, falling back to `BIGNUM` otherwise. This behavior can be overriden with macros. See the README and `src/ff3.c` for more info.

### Performance

These changes net a 6-7x speedup in the FF3 benchmarks over the current `BIGNUM` implementation on an M1 Macbook Pro (aarch64)

#### Before: `BIGNUM`-based Algorithm

```sh
FPE_CFLAGS='-DFPE_USE_BIGNUM' make benchmark && ./benchmark
```

```text
...
Benchmark: ff3 encrypt 10-character radix-10 input for 1000000 iterations...    7.233376 seconds
Benchmark: ff3 decrypt 10-character radix-10 input for 1000000 iterations...    7.267057 seconds
Benchmark: ff3 encrypt 10-character radix-62 input for 1000000 iterations...    7.086594 seconds
Benchmark: ff3 decrypt 10-character radix-62 input for 1000000 iterations...    7.637079 seconds
```

#### After: 128-bit Integer Based Algorithm

```sh
make benchmark && ./benchmark
```

```text
Benchmark: ff3 encrypt 10-character radix-10 input for 1000000 iterations...    1.209198 seconds
Benchmark: ff3 decrypt 10-character radix-10 input for 1000000 iterations...    1.205125 seconds
Benchmark: ff3 encrypt 10-character radix-62 input for 1000000 iterations...    1.223414 seconds
Benchmark: ff3 decrypt 10-character radix-62 input for 1000000 iterations...    1.219719 seconds
```